Move Shredder and TypeHierarchy to its own shredder package
