+++ 3-enrich/scala-common-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/common/utils/shredder/TypeHierarchy.scala
@@ -13,6 +13,7 @@
 package com.snowplowanalytics.snowplow.enrich
 package common
 package utils
+package shredder
 
 // Jackson
 import com.github.fge.jackson.JacksonUtils
