+++ 3-enrich/scala-common-enrich/src/test/scala/com.snowplowanalytics.snowplow.enrich.common/utils/shredder/ShredderSpec.scala
@@ -15,6 +15,7 @@ package snowplow
 package enrich
 package common
 package utils
+package shredder
 
 // Snowplow Utils
 import util.Tap._
