+++ 3-enrich/scala-common-enrich/src/test/scala/com.snowplowanalytics.snowplow.enrich.common/utils/shredder/TypeHierarchySpec.scala
@@ -13,6 +13,7 @@
 package com.snowplowanalytics.snowplow.enrich
 package common
 package utils
+package shredder
 
 // Specs2
 import org.specs2.Specification
